backend/requirements.txt
uvicorn
fastapi
fastapi-pagination
pytest
SQLalchemy
marshmallow-sqlalchemy
psycopg2-binary
alembic
python-jose
passlib
python-dotenv
colorama
cloudinary
python-multipart
